First working tree.py separation, with TODOs
