+++ src/gordion/repository.py
@@ -2,8 +2,6 @@ import os
 import subprocess
 import git
 import gordion
-from typing import List
-import shutil
 
 
 class Repository:
@@ -12,17 +10,11 @@ class Repository:
 
   """
 
-  def __init__(self, path: str, parent=None) -> None:
+  def __init__(self, path: str) -> None:
     self.path = path
     self.name = os.path.basename(self.path)
     self.url = ''
     self.fetched = False
-    self.parent: Repository = parent
-    self.children: dict[str, Repository] = {}
-    self.yeditor = gordion.YamlEditor(os.path.join(self.path, 'gordion.yaml'))
-
-    if Repository._exists(self.path):
-      self.ensure()
 
   def ensure(self, url: str = ''):
     """
@@ -38,14 +30,16 @@ class Repository:
       self.url = url
       if Repository._exists(self.path):
         if self.url != self.handle.remotes.origin.url:
-          self._check_different_repo_same_path(self._root())
-          gordion.Repository._safe_remove_repo(self.path)
+          pass
+          # TODO save these
+          # self._check_different_repo_same_path(self._root())
+          # gordion.Repository._safe_remove_repo(self.path)
 
     # Check for different repositories cloned to the same path.
-    self._check_different_repo_same_path(self._root())
+    # self._check_different_repo_same_path(self._root())
 
     # Check for a duplicates repository cloned at different paths.
-    self._check_duplicate_repo_path(self._root())
+    # self._check_duplicate_repo_path(self._root())
 
     # Clone if necessary. At this point the mirror should exist regardless of whether the repository
     # exists. so ensure it first.
@@ -57,54 +51,18 @@ class Repository:
 
     # Reload objects.
     self.handle = git.Repo(self.path)
-    self.yeditor.reload()
 
-  def _root(self):
-    """
-    Recursively returns the root repository object.
-    """
-    if self.parent:
-      return self.parent._root()
-    else:
-      return self
-
-  def _relpath(self) -> str:
-    return os.path.relpath(self.path, os.path.dirname(self._root().path))
-
-  def _listed_path(self) -> str:
-    listed_path = ''
-    if self.parent:
-      gpath = self.parent.yeditor.read_repository_gpath(self.name)
-      listed_path = f"{self.parent._relpath()} lists {gpath}"
-    else:
-      listed_path = f"{self._relpath()} (root)"
-
-    return listed_path
-
-  def update(self, tag: str, branch_name: str, force: bool = False) -> None:
+  def update(self, commit: git.Commit, branch_name: str, force: bool = False) -> None:
     """
     Updates the repository to the specified commit and optional branch, as long as information will
     not be lost in the process, otherwise it will raise descriptive errors about what to do next.
 
     """
 
-    commit: git.Commit = self._verify_tag(tag)
-
-    # Check for duplicate tag
-    root = self._root()
-    self._check_duplicate_repo_tag(tag, root)
-
     # If the commit does not change, we are done. Allow user to manually checkout a HEAD or
     # different branch name and still satisfy the update.
     if self.handle.head.commit.hexsha != commit.hexsha:
       self._update_moving_commit(commit, branch_name, force)
-
-    self.yeditor.reload()
-    self._update_children(branch_name, force)
-
-    # Cleanup detached repositories.
-    if self is root:
-      self._clean_detached_repos(force)
 
   def _update_moving_commit(self, commit: git.Commit, branch_name: str,
                             force: bool = False) -> None:
@@ -193,73 +151,6 @@ class Repository:
           self._check_dangling_commit(commit)
           self.handle.git.checkout(commit)
 
-  def _list_child_repository_paths(self) -> List[str]:
-    paths = []
-    for _, repo in self.children.items():
-      paths.append(repo.path)
-      paths.extend(repo._list_child_repository_paths())
-    return paths
-
-  @staticmethod
-  def _safe_remove_repo(path, force: bool = False):
-    assert gordion.Repository._exists(path)
-    repo = git.Repo(path)
-
-    # Check if repository has local changes.
-    if repo.is_dirty(untracked_files=True):
-      if not force:
-        raise gordion.UnsafeRemoveDirty(path)
-
-    # Check if any information would be lost from local branches if we delete this repository.
-    for local_branch in repo.branches:
-      # If there is a tracking branch, ensure the local branch is not ahead of it.
-      tracking_branch = local_branch.tracking_branch()
-      if tracking_branch:
-        merge_base = repo.merge_base(local_branch, tracking_branch)
-        commits_ahead = list(repo.iter_commits(
-            f'{merge_base[0].hexsha}..{local_branch.commit.hexsha}'))
-
-        if commits_ahead:
-          raise gordion.UnsafeRemoveLocalBranchAhead(path, local_branch.name,
-                                                     tracking_branch.name, len(commits_ahead))
-
-      # There is no tracking branch, so error.
-      else:
-        raise gordion.UnsafeRemoveLocalBranchNoTrackingBranch(path, local_branch.name)
-
-    # Error if the repository has stashes that will be lost by the deletion.
-    stashes = repo.git.stash('list')
-    if stashes:
-      raise gordion.UnsafeRemoveStashes(path, stashes)
-
-    # If we reach here, it's safe to delete the repository
-    print(f"Deleting directory: {path}")
-    shutil.rmtree(path)
-
-  def _clean_detached_repos(self, force: bool = False):
-    # Get all the paths of all repositories:
-    root = self._root()
-    child_paths = root._list_child_repository_paths()
-
-    # Delete git repositories.
-    for dirpath, dirnames, _ in os.walk(os.path.join(root.path, 'gordion'), topdown=True):
-      for dirname in dirnames:
-        full_dirpath = os.path.join(dirpath, dirname)
-        if (os.path.exists(full_dirpath) and not gordion.is_related_path(full_dirpath,
-                                                                         child_paths)):
-          if gordion.Repository._exists(full_dirpath):
-            gordion.Repository._safe_remove_repo(full_dirpath, force)
-
-    # Delete everything else that is not related to the gordion paths.
-    for dirpath, dirnames, _ in os.walk(os.path.join(root.path, 'gordion'), topdown=True):
-      for dirname in dirnames:
-        full_dirpath = os.path.join(dirpath, dirname)
-        if (os.path.exists(full_dirpath) and not gordion.is_related_path(full_dirpath,
-                                                                         child_paths)):
-          print(f"Deleting directory: {full_dirpath}")
-          assert not gordion.Repository._exists(full_dirpath)  # Removed above.
-          shutil.rmtree(full_dirpath)
-
   def _check_dangling_commit(self, commit):
     """
     Checks if the commit is dangling (does not belong to a branch) and raises an error if it is
@@ -327,22 +218,6 @@ class Repository:
     # Check against the other's children
     for _, other_child in other.children.items():
       Repository._check_duplicate_repo_tag(self, target_tag, other_child)
-
-  def _update_children(self, branch_name: str, force: bool):
-    root = self._root()
-    self.children = {}
-
-    # Open the gordion yaml file for this repository if it exists.
-    if self.yeditor.exists():
-      assert self.yeditor.yaml_data
-      for child_name, child_info in self.yeditor.yaml_data['repositories'].items():
-        # Create child repository objects
-        gpath = self.yeditor.read_repository_gpath(child_name)
-        child_path = os.path.join(root.path, 'gordion', gpath)
-        child = Repository(child_path, self)
-        child.ensure(child_info['url'])
-        child.update(child_info['tag'], branch_name, force)
-        self.children[child_name] = child
 
   def _verify_head_wont_be_lost(self, commit):
     """
